Map:is closed

+++ libft/ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: aheinane <aheinane@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/24 15:11:20 by aheinane          #+#    #+#             */
-/*   Updated: 2023/11/13 15:24:32 by aheinane         ###   ########.fr       */
+/*   Updated: 2024/09/26 13:16:53 by aheinane         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_strlen(const char *str )
+int	ft_strlen(const char *str )
 {
 	size_t	i;
 	size_t	count;
